created files for api routes for login and new user and new file for the frontend js for creating new user
