+++ controllers/api/userRoutes.js
@@ -1,0 +1,9 @@
+const router = require('express').Router();
+
+const {User} = require('../../models/')
+
+
+
+
+
+module.exports = router
