+++ controllers/index.js
@@ -1,7 +1,9 @@
 const router = require('express').Router();
 const homeRoutes = require('./homeRoutes');
+const apiRoutes = require('./api');
 
 router.use('/',homeRoutes)
+router.use('/api', apiRoutes);
