+++ models/index.js
@@ -1,0 +1,4 @@
+const User = require('./User');
+
+
+module.exports = { User };
