+++ views/login.handlebars
@@ -27,8 +27,4 @@
   </div>
 </div>
 
-
-
-
-
-{{!-- <script src="./js/login.js"></script> --}}
+<script src="./js/login.js"></script>
