+++ views/new-user.handlebars
@@ -29,6 +29,8 @@
 </div>
 </div>
 
+<script src="./js/create-user.js"></script>
+
